Beta Release
Consider this something of a beta release. There are some incomplete
functionalities - the points assignment system is inoperable, and the
rider/horse entry is also yet to be implemented. However, I've included
a binary zip in this commit if you want to download those to a Windows
environment and click around a little bit.  :horse: